delivery

+++ lms/filolinko/src/components/shared/File.jsx
@@ -1,14 +1,19 @@
 import { useEffect, useState } from "react";
 import { useRecoilState } from "recoil";
+
 import { filesDelete, filesReload } from "../../state/triggers";
 import { FileType } from "../../scripts/fileType";
 import { removeDocument } from "../../scripts/fireStore";
 import { removeFile } from "../../scripts/storage";
 import Icons from "../../data/Icons.json";
+import Modal from "./Modal";
+import EditFile from "./EditFile";
 
 export default function File({ hook }) {
+  const [modal, setShowModal] = useState(false);
+  const [item, id] = hook;
   const [reload, setReload] = useRecoilState(filesReload);
-  const [remove, toggleRemove] = useRecoilState(filesDelete);
+  const [remove] = useRecoilState(filesDelete);
   const [wiggle, setWiggle] = useState("");
 
   useEffect(() => {
@@ -18,26 +23,32 @@ export default function File({ hook }) {
       setWiggle("");
     }
   }, [remove]);
-  const [item, id] = hook;
 
   async function removeItem() {
-    await removeFile(item.ref);
     await removeDocument(`/courses/${id}/files`, item.id);
+    if (item.type !== "link") {
+      await removeFile(item.ref);
+    }
     setReload(!reload);
   }
 
   return (
-    <div className={`file ${wiggle}`}>
+    <div className={`item ${wiggle}`}>
       <span>
         {remove && (
-          <i onClick={(e) => removeItem(e)} className={Icons.removeSmall} />
+          <i onClick={() => removeItem()} className={Icons.removeSmall} />
         )}
       </span>
-      <i className={FileType(item.type)} />
-      <p>{item.ref}</p>
-      <a href={item.url} download>
-        Download
+      <i onClick={() => setShowModal(true)} className={FileType(item.type)} />
+      <p className="item-name">{item.name}</p>
+      <a href={item.url} download target="_blank">
+        {item.type === "link" ? "Open Link" : "Download"}
       </a>
+      <Modal hook={[modal, setShowModal]}>
+        <EditFile
+          hook={[item, modal, setShowModal, id, item.id, reload, setReload]}
+        />
+      </Modal>
     </div>
   );
 }
